+++ src/test/java/utils/CourseDataGenerator.java
@@ -2,6 +2,8 @@ package utils;
 
 import com.github.javafaker.Faker;
 
+import java.math.BigDecimal;
+
 public class CourseDataGenerator {
 
     private static final Faker faker = new Faker();
@@ -39,9 +41,11 @@ public class CourseDataGenerator {
                     .numberBetween(1, 100) + " hours";
     }
 
-    public static Float largePrice() {
-        return (float) faker.number()
-                            .randomNumber(8, true);
+    public static BigDecimal largePrice() {
+        return BigDecimal.valueOf(
+                faker.number()
+                     .randomNumber(8, true)
+        );
     }
 
     public static String validThumbnailUrl() {
